requirements.txt
# Oracle SQL 튜닝 자동화 파이프라인 의존성

# Oracle Database 연결
oracledb>=1.4.0

# 데이터 처리
pandas>=1.5.0

# Excel 파일 생성
openpyxl>=3.1.0

# SSH 접속 (트레이스 파일 수집용)
paramiko>=3.0.0

# HTML 템플릿 처리
Jinja2>=3.1.0

# 설정 파일 처리
PyYAML>=6.0

# 로깅 (표준 라이브러리지만 명시)
# logging - 기본 포함

# 파일 처리
# os, re, datetime - 기본 포함

# 추가 유틸리티 (선택적)
# colorama>=0.4.0  # 콘솔 색상 출력용
# tqdm>=4.64.0     # 프로그레스 바용